requirements.txt
colorama
selenium
Pillow
imagehash
pyyaml